Consensus: MOVEONLY: move pow.o to consensus dir
