+++ src/consensus/pow.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "pow.h"
+#include "consensus/pow.h"
 
 #include "arith_uint256.h"
 #include "chain.h"
