+++ src/miner.cpp
@@ -7,12 +7,12 @@
 
 #include "amount.h"
 #include "consensus/consensus.h"
+#include "consensus/pow.h"
 #include "consensus/validation.h"
 #include "hash.h"
 #include "main.h"
 #include "net.h"
 #include "policy/policy.h"
-#include "pow.h"
 #include "primitives/transaction.h"
 #include "timedata.h"
 #include "util.h"
