+++ src/rpcmining.cpp
@@ -6,13 +6,13 @@
 #include "amount.h"
 #include "chainparams.h"
 #include "consensus/consensus.h"
+#include "consensus/pow.h"
 #include "consensus/validation.h"
 #include "core_io.h"
 #include "init.h"
 #include "main.h"
 #include "miner.h"
 #include "net.h"
-#include "pow.h"
 #include "rpcserver.h"
 #include "util.h"
 #include "validationinterface.h"
