+++ src/test/pow_tests.cpp
@@ -2,10 +2,10 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "consensus/pow.h"
 #include "main.h"
-#include "pow.h"
-#include "util.h"
 #include "test/test_bitcoin.h"
+#include "util.h"
 
 #include <boost/test/unit_test.hpp>
