+++ src/txdb.cpp
@@ -6,9 +6,9 @@
 #include "txdb.h"
 
 #include "chainparams.h"
+#include "consensus/pow.h"
 #include "hash.h"
 #include "main.h"
-#include "pow.h"
 #include "uint256.h"
 
 #include <stdint.h>
